refactor: inline resolveContinueWatchingVideoId into HomeViewModelContinueWatching
Move the helper out of its own file per review feedback; keep the
function in the ViewModel file where all three call sites live. Trim
the KDoc to match the actual logic size.
